#270 changed GetKeptnBridgeURL to use K8s library, refactored to split up import cycle
