+++ pkg/adapter/evaluationfinished_adapter.go
@@ -2,11 +2,10 @@ package adapter
 
 import (
 	"fmt"
+	"github.com/keptn-contrib/dynatrace-service/pkg/credentials"
 	keptnapi "github.com/keptn/go-utils/pkg/api/utils"
 	keptnv2 "github.com/keptn/go-utils/pkg/lib/v0_2_0"
 	"os"
-
-	"github.com/keptn-contrib/dynatrace-service/pkg/common"
 )
 
 // EvaluationFinishedAdapter godoc
@@ -79,7 +78,7 @@ func (a EvaluationFinishedAdapter) GetTag() string {
 // GetLabels returns a map of labels
 func (a EvaluationFinishedAdapter) GetLabels() map[string]string {
 	labels := a.event.Labels
-	keptnBridgeURL, err := common.GetKeptnBridgeURL()
+	keptnBridgeURL, err := credentials.GetKeptnBridgeURL()
 	if labels == nil {
 		labels = make(map[string]string)
 	}
