+++ pkg/adapter/testtriggered_adapter.go
@@ -1,7 +1,7 @@
 package adapter
 
 import (
-	"github.com/keptn-contrib/dynatrace-service/pkg/common"
+	"github.com/keptn-contrib/dynatrace-service/pkg/credentials"
 	keptnv2 "github.com/keptn/go-utils/pkg/lib/v0_2_0"
 )
 
@@ -75,7 +75,7 @@ func (a TestTriggeredAdapter) GetTag() string {
 // GetLabels returns a map of labels
 func (a TestTriggeredAdapter) GetLabels() map[string]string {
 	labels := a.event.Labels
-	keptnBridgeURL, err := common.GetKeptnBridgeURL()
+	keptnBridgeURL, err := credentials.GetKeptnBridgeURL()
 	if labels == nil {
 		labels = make(map[string]string)
 	}
