+++ pkg/event_handler/action.go
@@ -24,7 +24,7 @@ type ActionHandler struct {
  * Retrieves Dynatrace Credential information
  */
 func (eh ActionHandler) GetDynatraceCredentials(keptnEvent adapter.EventContentAdapter) (*config.DynatraceConfigFile, *credentials.DTCredentials, error) {
-	dynatraceConfig, err := config.GetDynatraceConfig(keptnEvent, eh.Logger)
+	dynatraceConfig, err := adapter.GetDynatraceConfig(keptnEvent, eh.Logger)
 	if err != nil {
 		eh.Logger.Error("failed to load Dynatrace config: " + err.Error())
 		return nil, nil, err
@@ -75,7 +75,7 @@ func (eh ActionHandler) HandleEvent() error {
 			comment = comment + ": " + actionTriggeredData.Action.Description
 		}
 
-		dynatraceConfig, err := config.GetDynatraceConfig(keptnEvent, eh.Logger)
+		dynatraceConfig, err := adapter.GetDynatraceConfig(keptnEvent, eh.Logger)
 		if err != nil {
 			eh.Logger.Error("failed to load Dynatrace config: " + err.Error())
 			return err
